+++ libs/forms-shadcn/src/index.ts
@@ -10,6 +10,8 @@ export { FieldCheckbox } from './lib/fields/field-checkbox'
 export { FieldCombobox } from './lib/fields/field-combobox'
 export { FieldDate } from './lib/fields/field-date'
 export { FieldDateRange } from './lib/fields/field-date-range'
+export { FieldDateTimePicker } from './lib/fields/field-datetime-picker'
+export { FieldDuration } from './lib/fields/field-duration'
 export { FieldHidden, type HiddenFieldProps } from './lib/fields/field-hidden'
 export { FieldNativeSelect } from './lib/fields/field-native-select'
 export { FieldNumber } from './lib/fields/field-number'
@@ -33,6 +35,8 @@ export type {
   DateRangeFieldProps,
   DateRangePreset,
   DateRangeValue,
+  DateTimePickerFieldProps,
+  DurationFieldProps,
   NativeSelectFieldProps,
   NumberFieldProps,
   PasswordFieldProps,